[J:MEP-0] Move Oidc/Mock auth providers to nonWear/driven + restore internal
OidcAuthTokenProvider and MockAuthTokenProvider were placed in common/driven
during the variant-split refactor, which forced their visibility to be
widened from internal to public so the nonWear Koin module could consume
them. Architectural review showed these classes are used only by the
nonWear Koin binding — they belong in nonWear/driven. Moving them there
lets the visibility return to internal and confines OIDC client wiring
(OpenIdConnectClient, OIDC Ktor transport) to the variant that actually
needs it.

- git mv OidcAuthTokenProvider.kt, MockAuthTokenProvider.kt, LH.kt logger
  helper from common/driven/internal to nonWear/driven/internal.
- Update package declarations and imports to
  cz.adamec.timotej.snag.authentication.fe.nonwear.driven.internal
  (per "package follows path on renames" feedback).
- Restore `internal class OidcAuthTokenProvider` and
  `internal class MockAuthTokenProvider` visibility.
- Update AuthenticationDrivenNonWearModule imports to reference the new
  nonwear.driven.internal package.
- Move libs.oidc.ktor implementation dependency from common/driven to
  nonWear/driven (it was only needed for OpenIdConnectClient's Ktor
  transport used by OidcAuthTokenProvider).

common/driven now contains only platform-agnostic and variant-agnostic
infrastructure (Koin stub + expect/actual platform token-storage bindings
shared by both variants).

Verification: ./gradlew :feat:authentication:fe:common:driven:check
:feat:authentication:fe:nonWear:driven:check
:feat:authentication:fe:wear:driven:check green.
