+++ feat/authentication/fe/common/driven/build.gradle.kts
@@ -20,7 +20,6 @@ kotlin {
             implementation(projects.lib.configuration.fe.api)
             implementation(libs.oidc.appsupport)
             implementation(libs.oidc.tokenstore)
-            implementation(libs.oidc.ktor)
         }
         val webMain by getting {
             dependencies {
