+++ feat/authentication/fe/nonWear/driven/build.gradle.kts
@@ -21,6 +21,7 @@ kotlin {
             implementation(projects.lib.configuration.fe.api)
             implementation(libs.oidc.appsupport)
             implementation(libs.oidc.tokenstore)
+            implementation(libs.oidc.ktor)
         }
     }
 }
